refactor: split up the dragdrop actions into separate files
