+++ packages/dnd-core/src/actions/dragDrop/endDrag.ts
@@ -1,0 +1,30 @@
+declare var require: any
+
+import {
+	DragDropManager,
+	SentinelAction,
+	DragDropMonitor,
+} from '../../interfaces'
+import { END_DRAG } from './types'
+
+const invariant = require('invariant')
+
+export default function createEndDrag<Context>(
+	manager: DragDropManager<Context>,
+) {
+	return function endDrag(): SentinelAction {
+		const monitor = manager.getMonitor()
+		const registry = manager.getRegistry()
+		verifyIsDragging(monitor)
+
+		const sourceId = monitor.getSourceId()
+		const source = registry.getSource(sourceId!, true)
+		source.endDrag(monitor, sourceId!)
+		registry.unpinSource()
+		return { type: END_DRAG }
+	}
+}
+
+function verifyIsDragging(monitor: DragDropMonitor) {
+	invariant(monitor.isDragging(), 'Cannot call endDrag while not dragging.')
+}
