+++ packages/dnd-core/src/actions/dragDrop/types.ts
@@ -1,0 +1,6 @@
+export const INIT_COORDS = 'dnd-core/INIT_COORDS'
+export const BEGIN_DRAG = 'dnd-core/BEGIN_DRAG'
+export const PUBLISH_DRAG_SOURCE = 'dnd-core/PUBLISH_DRAG_SOURCE'
+export const HOVER = 'dnd-core/HOVER'
+export const DROP = 'dnd-core/DROP'
+export const END_DRAG = 'dnd-core/END_DRAG'
